updated firebase lib and token access code changes with new method
