+++ android-sdk/build.gradle
@@ -60,13 +60,13 @@ dependencies {
 
     compile 'com.android.support:appcompat-v7:27.0.2'
     // for Android Advertising Id
-    compile 'com.google.android.gms:play-services-ads:11.6.2'
+    compile 'com.google.android.gms:play-services-ads:15.0.1'
     // for JSON parsing
     compile 'com.google.code.gson:gson:2.8.0'
     // for annotations used (ex @NotNull)
     compile 'org.jetbrains:annotations:13.0'
     // for FCM
-    compile 'com.google.firebase:firebase-messaging:11.6.2'
+    compile 'com.google.firebase:firebase-messaging:17.3.1'
 
     testCompile 'junit:junit:4.12'
 }
